- composite REST end point to serve composite REST objects
